moved _handle_rating_tags and _resolve_conflicting_ratings from FileSystemPlayer to FileSystemProvider
